IA-16: separate "any segment register" and "%es" constraints
Before this patch, there was just one assembly operand
constraint "Q" which could mean either "any (available)
segment register" or "the %es register".

This patch adds a new constraint "e".

The intent is that the user should specify "e" if an operand
absolutely must go into %es --- for example because an ABI
mandates it --- and specify "Q" if it is fine for it to go
into (say) %ds, if %ds ever becomes available for allocation.

"e" currently maps internally to the same register class as
"Q", namely SEGMENT_REGS, but this may change.

I have also added two test programs which check that both
"Q" and "e" work as they should.