IA-16: bug fix: peephole on `movw %es, %bx; movw (%bx), %bx'
One of the peephole rules was wrongly optimizing `movw %es,
%bx; movw mem, %bx' into `movw mem, %bx' even when mem was
`(%bx)'.  The correct thing to do was to check %bx for
liveness before the `movw mem, ...'

This fixes a regression in gcc/testsuite/gcc.c-torture/
execute/bitfld-3.c.  The regression showed up
intermittently, so I initially misdiagnosed it as being
caused by changes to the register file.