IA-16: flag "does not support far static" error at pt. of decl.
Previously the error message for "tiny code model does not
support '__far' static storage variables" gave a source line
number corresponding to to the end of the program being
compiled, which is not very useful.  The error message now
shows the line number for the offending declaration.

I have also updated the test cases accordingly.